NIFI-14727 Refactor RecordBuffer interface
