+++ nifi-extension-bundles/nifi-aws-bundle/nifi-aws-kinesis/src/main/java/org/apache/nifi/processors/aws/kinesis/ConsumeKinesis.java
@@ -48,8 +48,6 @@ import org.apache.nifi.proxy.ProxySpec;
 import org.apache.nifi.serialization.RecordReaderFactory;
 import org.apache.nifi.serialization.RecordSetWriterFactory;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.awscore.AwsClient;
-import software.amazon.awssdk.awscore.client.builder.AwsAsyncClientBuilder;
 import software.amazon.awssdk.http.Protocol;
 import software.amazon.awssdk.http.async.SdkAsyncHttpClient;
 import software.amazon.awssdk.http.nio.netty.Http2Configuration;
@@ -84,21 +82,26 @@ import java.util.Set;
 import java.util.UUID;
 
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.APPROXIMATE_ARRIVAL_TIMESTAMP;
+import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.MIME_TYPE;
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.PARTITION_KEY;
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.RECORD_COUNT;
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.RECORD_ERROR_MESSAGE;
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.SEQUENCE_NUMBER;
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.SHARD_ID;
+import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.STREAM_NAME;
 import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.SUB_SEQUENCE_NUMBER;
 
 @InputRequirement(InputRequirement.Requirement.INPUT_FORBIDDEN)
 @Tags({"amazon", "aws", "kinesis", "consume", "stream", "record"})
 @CapabilityDescription("Consumes data from the specified AWS Kinesis stream and outputs a FlowFile for every processed Record (raw) " +
         "or a FlowFile for a batch of processed records if a Record Reader and Record Writer are configured. " +
-        "AWS Kinesis Client Library can take several seconds to initialise before starting to fetch data. " +
+        "AWS Kinesis Client Library can take a few minutes on the first start and several seconds on subsequent start " +
+        "to initialise before starting to fetch data. " +
         "Uses DynamoDB for check pointing and coordination, and CloudWatch (optional) for metrics. " +
         "Ensure that the credentials provided have access to DynamoDB and CloudWatch (optional) along with Kinesis.")
 @WritesAttributes({
+        @WritesAttribute(attribute = STREAM_NAME,
+                description = "The name of the Kinesis Stream from which all Kinesis Records in the Flow File were read"),
         @WritesAttribute(attribute = SHARD_ID,
                 description = "Shard ID from which all Kinesis Records in the Flow File were read"),
         @WritesAttribute(attribute = PARTITION_KEY,
@@ -109,7 +112,7 @@ import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.SU
                 description = "A SubSequence Number of the last Kinesis Record in the Flow File. Generated by KPL when aggregating records into a single Kinesis Record"),
         @WritesAttribute(attribute = APPROXIMATE_ARRIVAL_TIMESTAMP,
                 description = "Approximate arrival timestamp of the last Kinesis Record in the Flow File"),
-        @WritesAttribute(attribute = "mime.type",
+        @WritesAttribute(attribute = MIME_TYPE,
                 description = "Sets the mime.type attribute to the MIME Type specified by the Record Writer (if configured)"),
         @WritesAttribute(attribute = RECORD_COUNT,
                 description = "Number of records written to the FlowFiles by the Record Writer (if configured)"),
@@ -124,7 +127,8 @@ import static org.apache.nifi.processors.aws.kinesis.ConsumeKinesisAttributes.SU
 @SystemResourceConsideration(resource = SystemResource.NETWORK, description = "Kinesis Client Library will continually poll for new Records, " +
         "requesting up to a maximum number of Records/bytes per call. This can result in sustained network usage.")
 @SystemResourceConsideration(resource = SystemResource.MEMORY, description = "ConsumeKinesis buffers Kinesis Records in memory until they can be processed. " +
-        "The maximum size of the buffer is controlled by the 'Max Bytes to Buffer' property.")
+        "The maximum size of the buffer is controlled by the 'Max Bytes to Buffer' property. " +
+        "In addition, Kinesis Client Library may cache some amount of data for each shard when the processor's buffer is full.")
 public class ConsumeKinesis extends AbstractProcessor {
 
     private static final Duration CONNECTION_TIMEOUT = Duration.ofSeconds(30);
@@ -264,7 +268,9 @@ public class ConsumeKinesis extends AbstractProcessor {
     private volatile KinesisAsyncClient kinesisClient;
     private volatile Scheduler kinesisScheduler;
 
-    private volatile RecordBuffer recordBuffer;
+    private volatile String streamName;
+    private volatile RecordBuffer.ForProcessor recordBuffer;
+
     private volatile boolean useReader = false;
     private volatile Optional<ReaderRecordProcessor> readerRecordProcessor = Optional.empty();
 
@@ -318,14 +324,15 @@ public class ConsumeKinesis extends AbstractProcessor {
                 .httpClient(createHttpClientBuilder(context).build())
                 .build();
 
-        final String streamName = context.getProperty(KINESIS_STREAM_NAME).getValue();
+        streamName = context.getProperty(KINESIS_STREAM_NAME).getValue();
         final InitialPositionInStreamExtended initialPositionExtended = getInitialPosition(context);
         final SingleStreamTracker streamTracker = new SingleStreamTracker(streamName, initialPositionExtended);
 
         final long maxBytesToBuffer = context.getProperty(MAX_BYTES_TO_BUFFER).asDataSize(DataUnit.B).longValue();
         final Duration checkpointInterval = context.getProperty(CHECKPOINT_INTERVAL).asDuration();
-        recordBuffer = new RecordBuffer(getLogger(), maxBytesToBuffer, checkpointInterval);
-        final ShardRecordProcessorFactory recordProcessorFactory = () -> new ConsumeKinesisRecordProcessor(recordBuffer);
+        final MemoryBoundRecordBuffer memoryBoundRecordBuffer = new MemoryBoundRecordBuffer(getLogger(), maxBytesToBuffer, checkpointInterval);
+        recordBuffer = memoryBoundRecordBuffer;
+        final ShardRecordProcessorFactory recordProcessorFactory = () -> new ConsumeKinesisRecordProcessor(memoryBoundRecordBuffer);
 
         final String applicationName = context.getProperty(APPLICATION_NAME).getValue();
         final String workerId = UUID.randomUUID().toString();
@@ -476,10 +483,10 @@ public class ConsumeKinesis extends AbstractProcessor {
         }
     }
 
-    private static void processRecordsAsRaw(final ProcessSession session, final String shardId, final List<KinesisClientRecord> records) {
+    private void processRecordsAsRaw(final ProcessSession session, final String shardId, final List<KinesisClientRecord> records) {
         for (final KinesisClientRecord record : records) {
             FlowFile flowFile = session.create();
-            flowFile = session.putAllAttributes(flowFile, ConsumeKinesisAttributes.fromKinesisRecord(shardId, record));
+            flowFile = session.putAllAttributes(flowFile, ConsumeKinesisAttributes.fromKinesisRecord(streamName, shardId, record));
 
             flowFile = session.write(flowFile, out -> Channels.newChannel(out).write(record.data()));
 
@@ -487,8 +494,8 @@ public class ConsumeKinesis extends AbstractProcessor {
         }
     }
 
-    private static void processRecordsWithReader(final ReaderRecordProcessor recordProcessor, final ProcessSession session, final String shardId, final List<KinesisClientRecord> records) {
-        final ProcessingResult result = recordProcessor.processRecords(session, shardId, records);
+    private void processRecordsWithReader(final ReaderRecordProcessor recordProcessor, final ProcessSession session, final String shardId, final List<KinesisClientRecord> records) {
+        final ProcessingResult result = recordProcessor.processRecords(session, streamName, shardId, records);
 
         session.transfer(result.successFlowFiles(), REL_SUCCESS);
         session.transfer(result.parseFailureFlowFiles(), REL_PARSE_FAILURE);
@@ -499,10 +506,10 @@ public class ConsumeKinesis extends AbstractProcessor {
      */
     private static class ConsumeKinesisRecordProcessor implements ShardRecordProcessor {
 
-        private final RecordBuffer recordBuffer;
+        private final RecordBuffer.ForKinesisClientLibrary recordBuffer;
         private volatile @Nullable ShardBufferId bufferId;
 
-        ConsumeKinesisRecordProcessor(final RecordBuffer recordBuffer) {
+        ConsumeKinesisRecordProcessor(final MemoryBoundRecordBuffer recordBuffer) {
             this.recordBuffer = recordBuffer;
         }
